Expired sessions are automatically refreshed
How this is implemented:
1. When user logs in, the client cookie \"autoLogin\" is set to \"1\".
2. When user has not visited Poggit for a long time such that the
   session expired, the session cookie (previously PHPSESSID, now
   changed to POGGIT) is no longer valid because PHP cannot identify its
   value. Therefore, the user is interpreted as "logged out" (having an
   invalid cookie, hence a new one is generated by PHP).
3. However, the user still has the \"autoLogin\" set, which did not
   expire. As the server receives this cookie, the server knows that the
   user was previously logged in but the session expired, so it tries to
   authenticate the user again by sending the user to GitHub.

This may worsen the user's landing experience as the redirects in the
login flow sometimes take several seconds. However, this is unlikely to
cause user loss as the user has already logged in (so this is not the
first impression), and this does not affect every single user session.